SX tailwind: remove '@babel/register' call from NPM
This should remove the `@babel/register` from NPM version since it's needed only for Adeira Universe monorepo and it's problematic when exported outside of the repo.

I tried to use the modified NPM output in SX Tailwind website and it worked well. It also works well under Universe workspace. This issue should be eventually solved by design with the migration to Bazel.

adeira-source-id: ec0888da1d211ac91367d7062b5d18de8edd53cf