Deprecate '@adeira/test-utils'
I was thinking for a while what to do with `@adeira/test-utils` package. There are only two functions (`evaluateGraphQLResolver` and `generateTestsFromFixtures`) the GraphQL resolver evaluation seems to be pretty useless/outdated. On the other hand the fixtures test generator is quite popular and helpful.

I'd like to propose a deprecation of https://www.npmjs.com/package/@adeira/test-utils and I'd like to create a new NPM package instead which is specializing only on the fixtures testing.

adeira-source-id: 214ec1ccb0e91b58d003fb761dd74ac880dad64b